+++ CHANGELOG.md
@@ -10,5 +10,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Use Edit Instance button for all instances
+- Use new Microsoft account login endpoint and add entitlement checks back in
 
 ### Misc